+++ pkg/sdk/poc/generator/parameter_builders.go
@@ -1,45 +1,54 @@
 package generator
 
-func (v *queryStruct) Assignment(name string, kind string, transformer *ParameterTransformer) *queryStruct {
+func (v *queryStruct) assignment(name string, kind string, transformer *ParameterTransformer) *queryStruct {
 	v.fields = append(v.fields, NewField(name, kind, Tags().Parameter(), transformer))
 	return v
 }
 
-func (v *queryStruct) ListAssignment(sqlPrefix string, listItemKind string, transformer *ParameterTransformer) *queryStruct {
+func (v *queryStruct) Assignment(sqlPrefix string, kind string, transformer *ParameterTransformer) *queryStruct {
 	if transformer != nil {
 		transformer = transformer.SQL(sqlPrefix)
 	} else {
 		transformer = ParameterOptions().SQL(sqlPrefix)
 	}
-	return v.Assignment(sqlToFieldName(sqlPrefix, true), KindOfSlice(listItemKind), transformer)
+	return v.assignment(sqlToFieldName(sqlPrefix, true), kind, transformer)
+}
+
+func (v *queryStruct) OptionalAssignment(sqlPrefix string, kind string, transformer *ParameterTransformer) *queryStruct {
+	if len(kind) > 0 && kind[0] != '*' {
+		kind = KindOfPointer(kind)
+	}
+	return v.Assignment(sqlPrefix, kind, transformer)
+}
+
+func (v *queryStruct) ListAssignment(sqlPrefix string, listItemKind string, transformer *ParameterTransformer) *queryStruct {
+	return v.Assignment(sqlPrefix, KindOfSlice(listItemKind), transformer)
+}
+
+func (v *queryStruct) NumberAssignment(sqlPrefix string, transformer *ParameterTransformer) *queryStruct {
+	return v.Assignment(sqlPrefix, "int", transformer)
+}
+
+func (v *queryStruct) OptionalNumberAssignment(sqlPrefix string, transformer *ParameterTransformer) *queryStruct {
+	return v.Assignment(sqlPrefix, "*int", transformer)
 }
 
 func (v *queryStruct) TextAssignment(sqlPrefix string, transformer *ParameterTransformer) *queryStruct {
-	if transformer != nil {
-		transformer = transformer.SQL(sqlPrefix)
-	} else {
-		transformer = ParameterOptions().SQL(sqlPrefix)
-	}
-	return v.Assignment(sqlToFieldName(sqlPrefix, true), "string", transformer)
+	return v.Assignment(sqlPrefix, "string", transformer)
 }
 
 func (v *queryStruct) OptionalTextAssignment(sqlPrefix string, transformer *ParameterTransformer) *queryStruct {
-	if transformer != nil {
-		transformer = transformer.SQL(sqlPrefix)
-	} else {
-		transformer = ParameterOptions().SQL(sqlPrefix)
-	}
-	return v.Assignment(sqlToFieldName(sqlPrefix, true), "*string", transformer)
+	return v.Assignment(sqlPrefix, "*string", transformer)
+}
+
+func (v *queryStruct) BooleanAssignment(sqlPrefix string, transformer *ParameterTransformer) *queryStruct {
+	return v.Assignment(sqlPrefix, "bool", transformer)
+}
+
+func (v *queryStruct) OptionalBooleanAssignment(sqlPrefix string, transformer *ParameterTransformer) *queryStruct {
+	return v.Assignment(sqlPrefix, "*bool", transformer)
 }
 
 func (v *queryStruct) OptionalIdentifierAssignment(sqlPrefix string, identifierKind string, transformer *ParameterTransformer) *queryStruct {
-	if transformer != nil {
-		transformer = transformer.SQL(sqlPrefix)
-	} else {
-		transformer = ParameterOptions().SQL(sqlPrefix)
-	}
-	if len(identifierKind) > 0 && identifierKind[0] != '*' {
-		identifierKind = KindOfPointer(identifierKind)
-	}
-	return v.Assignment(sqlToFieldName(sqlPrefix, true), identifierKind, transformer)
+	return v.OptionalAssignment(sqlPrefix, identifierKind, transformer)
 }